vc: convert to utf8 only once

+++ avbuild.sh
@@ -982,7 +982,7 @@ config1(){
 " "$FFSRC_TOOLS/cmdutils.c"
     fi
     if $VC_BUILD; then # check ffmpeg version?
-      if [ "${VisualStudioVersion:0:2}" -gt 14 ] && `echo $LANG |grep -q zh`; then  # check ffmpeg version?
+      if [ "${VisualStudioVersion:0:2}" -gt 14 ] && `echo $LANG |grep -q zh` && [ ! -f config-utf8.h ] ; then  # check ffmpeg version?
         iconv -t "UTF-8" -f "GBK" config.h > config-utf8.h
         cp -f config{-utf8,}.h
       fi
